enable adding dependencies for workspace
Since the flutter or dart for adding pub dependencies doesn't work with workspace option, while it's better to add dependencies with the command, just added the package name, not clarifying the version. After making the package, it should be fine to run `melos bs`, if the package version is specified in melos.yaml or the corresponding pubspec.yaml.